after customer has entered info, creat customer is fired and current object is updated appropritately
