+++ lib/createCustomer.js
@@ -1,53 +1,5 @@
 'use strict';
 
-module.exports.createCustomer = ()=>{
-	 prompt.message = ''
-  prompt.delimiter = ` 
->`;
-  prompt.start();
-
-	let createCustomerForm = {
-	      name: {
-	        description: 'Enter Customer name (1/6)'
-	      },
-	      address: {
-	        description: 'Enter street address (2/6)'
-	      },
-	      city: {
-	        description: 'Enter city (3/6)'
-	      },
-	      state: {
-	        description: 'Enter state (4/6)'
-	      },
-	      postal_code: {
-	        description: 'Enter postal code (5/6)'
-	      },
-	      phone_number: {
-	        description: 'Enter phone number (6/6)'
-	      }
-	}
-
-	  prompt.get({
-	    properties: createCustomerForm
-	  }, function (err, result) {
-	    console.log('  name: ' + result.name);
-	    console.log('  address: ' + result.address);
-	    console.log('  city: ' + result.city);
-	    console.log('  state: ' + result.state);
-	    console.log('  postal code: ' + result.postal_code);
-	    console.log('  phone number: ' + result.phone_number);
-	    console.log(testing)
-
-	    prompt.start();
-	      prompt.get({
-	      properties: {
-	        confirm: {
-	          description: 'Look Ok? (y/n)'
-	        }
-	      }
-	    }, function (err, go) {
-	     	
-	    });
-
-  });
+module.exports.createCustomer = (arg) => {
+	console.log(arg)
 };
