+++ lib/currentObj.js
@@ -16,19 +16,16 @@ let currentObj = {
   checkoutInProgress: false,
   currentPage: "mainmenu",
   customerPath: ""
-
 }
-
-
-
 
 //takes two arguments
   //this first being the property to update
   //the second being the value to update it to.
-module.exports.updateCurrentObj = (args)=>{
-  if(args.length === 0 ){throw new Error ("There is no valid entry")};
-  currentObj[args[0]] = args[1];
+module.exports.updateCurrentObj = (arg1, arg2)=>{
+  if(arg1.length === 0 ){throw new Error ("There is no valid entry")};
+  currentObj[arg1] = arg2;
 };
+
 module.exports.clearCurrentObj = ()=>{
   currentObj = {
     customerid: "",
@@ -50,6 +47,5 @@ module.exports.clearCurrentObj = ()=>{
 };
 
 module.exports.getCurrentObj = ()=>{
-
   return currentObj;
 };
