+++ lib/pageHandler.js
@@ -1,15 +1,20 @@
 'use strict';
 
 const { getCurrentObj, updateCurrentObj } = require('./currentObj.js')
+const { productPopularity } = require('./productPopularity.js')
+const { createCustomerMenu } = require('./createCustomerMenu.js')
+// const {  }
+const currentObj = getCurrentObj()
 
 module.exports.pageHandler = (arg)=>{
   if(arg.length === 0 ){throw new Error ("There is no valid entry")};
   	switch (arg) {
 	  case '1':
-		console.log(arg)
+	  	// Show Create Customer Screen
+		createCustomerMenu()
 		break
 	  case '2':
-		console.log('2')
+		// Show Select Existing Customer Screen
 		break
 	  case '3':
 		console.log(arg)
@@ -24,7 +29,7 @@ module.exports.pageHandler = (arg)=>{
 		console.log(arg)
 		break
 	  case '7':
-	  	console.log(arg)
+	  	productPopularity()
 	  	break
 	}
   return {};
