+++ test/createCustomerMenu-test.js
@@ -1,0 +1,8 @@
+const {assert : {isFunction}} =require('chai');
+const {createCustomerMenu} = require('../lib/createCustomerMenu.js');
+
+describe ('createCustomer',()=>{
+  it('should be a function',()=>{
+    isFunction(createCustomerMenu)
+  });
+});
