+++ test/mainmenu-test.js
@@ -5,8 +5,5 @@ describe ('mainmenu',()=>{
   it('should be a function',()=>{
     isFunction(mainMenu)
   });
-  it('should have argument ',()=>{
-    throws(()=>{mainMenu()})//doesn't have an argument
-  });
 
 });
